requirements.txt
python-telegram-bot==20.0
feedparser==6.0.10
APScheduler==3.10.4